Redefine the error response
After inadvertedly causing an error we found it wasn't unmarshalling
properly. We then inspected the raw reply and found that the structure
of the error object was different than what was defined on the spec.
We have simply corrected that.

In case a per-status code approach needs to be taken section Response
Objects [0] of the 3.0.2 OpenAPI spec might prove useful.

0 - https://spec.openapis.org/oas/v3.0.2#responses-object